move op storage tests into new crate reth-optimism-storage
