+++ Cargo.toml
@@ -79,6 +79,7 @@ members = [
     "crates/optimism/payload/",
     "crates/optimism/primitives/",
     "crates/optimism/rpc/",
+    "crates/optimism/storage",
     "crates/payload/basic/",
     "crates/payload/builder/",
     "crates/payload/primitives/",
@@ -368,6 +369,7 @@ reth-optimism-forks = { path = "crates/optimism/hardforks" }
 reth-optimism-payload-builder = { path = "crates/optimism/payload" }
 reth-optimism-primitives = { path = "crates/optimism/primitives" }
 reth-optimism-rpc = { path = "crates/optimism/rpc" }
+reth-optimism-storage = { path = "crates/optimism/storage" }
 reth-payload-builder = { path = "crates/payload/builder" }
 reth-payload-primitives = { path = "crates/payload/primitives" }
 reth-payload-validator = { path = "crates/payload/validator" }
