+++ crates/optimism/storage/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "reth-optimism-storage"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+reth-primitives.workspace = true
+
+[dev-dependencies]
+reth-db-api.workspace = true
+reth-prune-types.workspace = true
+reth-stages-types.workspace = true
+
+[features]
+optimism = ["reth-primitives/optimism"]
