+++ crates/storage/db-api/src/models/mod.rs
@@ -372,60 +372,4 @@ mod tests {
         validate_bitflag_backwards_compat!(Withdrawals, UnusedBits::Zero);
         validate_bitflag_backwards_compat!(Requests, UnusedBits::Zero);
     }
-
-    #[cfg(feature = "optimism")]
-    #[test]
-    fn test_ensure_backwards_compatibility() {
-        assert_eq!(Account::bitflag_encoded_bytes(), 2);
-        assert_eq!(AccountHashingCheckpoint::bitflag_encoded_bytes(), 1);
-        assert_eq!(CheckpointBlockRange::bitflag_encoded_bytes(), 1);
-        assert_eq!(CompactClientVersion::bitflag_encoded_bytes(), 0);
-        assert_eq!(CompactU256::bitflag_encoded_bytes(), 1);
-        assert_eq!(CompactU64::bitflag_encoded_bytes(), 1);
-        assert_eq!(EntitiesCheckpoint::bitflag_encoded_bytes(), 1);
-        assert_eq!(ExecutionCheckpoint::bitflag_encoded_bytes(), 0);
-        assert_eq!(HeadersCheckpoint::bitflag_encoded_bytes(), 0);
-        assert_eq!(IndexHistoryCheckpoint::bitflag_encoded_bytes(), 0);
-        assert_eq!(PruneCheckpoint::bitflag_encoded_bytes(), 1);
-        assert_eq!(PruneMode::bitflag_encoded_bytes(), 1);
-        assert_eq!(PruneSegment::bitflag_encoded_bytes(), 1);
-        assert_eq!(Receipt::bitflag_encoded_bytes(), 2);
-        assert_eq!(ReceiptWithBloom::bitflag_encoded_bytes(), 0);
-        assert_eq!(SealedHeader::bitflag_encoded_bytes(), 0);
-        assert_eq!(StageCheckpoint::bitflag_encoded_bytes(), 1);
-        assert_eq!(StageUnitCheckpoint::bitflag_encoded_bytes(), 1);
-        assert_eq!(StoredBlockBodyIndices::bitflag_encoded_bytes(), 1);
-        assert_eq!(StoredBlockOmmers::bitflag_encoded_bytes(), 0);
-        assert_eq!(StoredBlockWithdrawals::bitflag_encoded_bytes(), 0);
-        assert_eq!(StorageHashingCheckpoint::bitflag_encoded_bytes(), 1);
-        assert_eq!(Withdrawals::bitflag_encoded_bytes(), 0);
-
-        // In case of failure, refer to the documentation of the
-        // [`validate_bitflag_backwards_compat`] macro for detailed instructions on handling
-        // it.
-        validate_bitflag_backwards_compat!(Account, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(AccountHashingCheckpoint, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(CheckpointBlockRange, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(CompactClientVersion, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(CompactU256, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(CompactU64, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(EntitiesCheckpoint, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(ExecutionCheckpoint, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(HeadersCheckpoint, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(IndexHistoryCheckpoint, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(PruneCheckpoint, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(PruneMode, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(PruneSegment, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(Receipt, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(ReceiptWithBloom, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(SealedHeader, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(StageCheckpoint, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(StageUnitCheckpoint, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(StoredBlockBodyIndices, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(StoredBlockOmmers, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(StoredBlockWithdrawals, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(StorageHashingCheckpoint, UnusedBits::NotZero);
-        validate_bitflag_backwards_compat!(Withdrawals, UnusedBits::Zero);
-        validate_bitflag_backwards_compat!(Requests, UnusedBits::Zero);
-    }
 }
